fix(admin): 仪表盘坏 SQL 与硬编码营收——eo_user 列名修正，OrderStats 扩营收字段走 eo_payment 汇总，状态计数合并单条 GROUP BY

+++ easyorange-backend/easyorange-admin/src/main/java/com/cartethyia/easyorange/admin/adapter/inbound/web/dto/response/DashboardStatsResponse.java
@@ -1,5 +1,6 @@
 package com.cartethyia.easyorange.admin.adapter.inbound.web.dto.response;
 
+import java.math.BigDecimal;
 import lombok.Builder;
 import lombok.Data;
 
@@ -19,7 +20,7 @@ public class DashboardStatsResponse {
 
     private Long todayOrders;
 
-    private Long totalRevenue;
+    private BigDecimal totalRevenue;
 
     private Long pendingReports;
 }
